+++ rumqttd/src/server/broker.rs
@@ -32,7 +32,6 @@ use crate::{Config, ConnectionId, ServerSettings};
 use tokio::net::{TcpListener, TcpStream};
 use tokio::time::error::Elapsed;
 use tokio::{task, time};
-use tracing::debug;
 
 #[derive(Debug, thiserror::Error)]
 #[error("Acceptor error")]
@@ -248,7 +247,6 @@ impl Broker {
                         }
                         _ => panic!("We only request for router metrics"),
                     }
-                    debug!("Prometheus update: {:?}", &metrics);
                     std::thread::sleep(Duration::from_secs(timeout));
                 }
             })?;
@@ -367,7 +365,7 @@ impl<P: Protocol + Clone + Send + 'static> Server<P> {
                 ),
                 LinkType::Remote => task::spawn(
                     remote(config, tenant_id, router_tx, network, protocol).instrument(
-                        tracing::info_span!(
+                        tracing::error_span!(
                             "remote_link",
                             client_id = field::Empty,
                             connection_id = field::Empty